* 📝 docs(README.md): remove commented out section on Git hook The section on Git hook was previously commented out, but it is now included in the README.md file.
* 🐛 fix(package.json): change cli file extension from .mjs to .cjs
* 🚀 chore(package.json): remove unnecessary banner from build script
The cli file extension has been changed from .mjs to .cjs to ensure compatibility with Node.js versions that do not support ECMAScript modules. The banner in the build script has been removed as it is no longer necessary with the change in file extension.

* 🐛 fix(githook.ts): change HOOK_PATH to HOOK_URL
The HOOK_PATH variable was changed to HOOK_URL to fix an issue where the symlink was not being created correctly. The HOOK_URL variable now uses the __filename global variable to get the path of the current file.

* 🐛 fix(tsconfig.json): change target to ES2020 and module to CommonJS
The target version of JavaScript has been changed to ES2020 to take advantage of the latest features of the language. The module has been changed to CommonJS to ensure compatibility with Node.js.