chore(config.ts): add OCO_GITPUSH environment variable to control git push behavior
The OCO_GITPUSH environment variable is added to the config file to
allow controlling the behavior of git push. If the variable is set to
'false', git push will be disabled. Otherwise, it will be enabled. This
provides flexibility in scenarios where git push is not desired, such as
during development or testing.