🔥 chore: remove .npmignore to streamline package publishing 🎉 feat: introduce .yamllint.yaml for YAML linting configuration 🔧 chore(package.json): add 'build:tsup' script for building with tsup ✨ feat: add tsup.config.ts for custom tsup build configuration
Removing the .npmignore file simplifies the package publishing process by relying on the "files" field in package.json or the default npm behavior. The addition of .yamllint.yaml provides a way to enforce style consistency in YAML files. The new 'build:tsup' script in package.json allows for an alternative build process using tsup, which is a faster and simpler bundler compared to other tools. The tsup.config.ts file is introduced to customize the tsup build process, including copying necessary WebAssembly files after a successful build, ensuring that the distribution package is complete and ready for use.